\name{Rblpapi.utils-package}
\alias{Rblpapi.utils-package}
\alias{Rblpapi.utils}
\docType{package}
\title{
\packageTitle{Rblpapi.utils}
}
\description{
\packageDescription{Rblpapi.utils}
}
\details{

  Several functions for data retrieval.
  

}
\author{
\packageAuthor{Rblpapi.utils}

Maintainer: \packageMaintainer{Rblpapi.utils}
}
%% \references{
%%   ~~ Literature or other references for background information ~~
%% }
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
## TODO
}
